requirements.txt
uvicorn~=0.20.0
fastapi~=0.89.1
pymongo~=4.3.3
pydantic~=1.10.4
psutil~=5.9.4
PyJWT~=2.6.0
passlib~=1.7.4
starlette~=0.22.0